[performance] 修改实体对于金钱的处理类型，修复每日表现下载bug

+++ perfect-service/src/main/java/com/perfect/entity/AccountReportEntity.java
@@ -4,6 +4,7 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import static com.perfect.mongodb.utils.EntityConstants.*;
@@ -35,13 +36,13 @@ public class AccountReportEntity implements Comparable<AccountReportEntity>{
     private Double pcCtr;     //PC点击率=点击次数/展现次数
 
     @Field(value = "pccost")
-    private Double pcCost;        //PC消费
+    private BigDecimal pcCost;        //PC消费
 
     @Field(value = "pccpc")
-    private Double pcCpc;     //PC平均点击价格=消费/点击次数
+    private BigDecimal pcCpc;     //PC平均点击价格=消费/点击次数
 
     @Field(value = "pccpm")
-    private Double pcCpm;       //PC千次展现消费
+    private BigDecimal pcCpm;       //PC千次展现消费
 
     @Field(value = "pccs")
     private Double pcConversion;      //PC转化
@@ -56,13 +57,13 @@ public class AccountReportEntity implements Comparable<AccountReportEntity>{
     private Double mobileCtr;
 
     @Field(value = "mcost")
-    private Double mobileCost;
+    private BigDecimal mobileCost;
 
     @Field(value = "mcpc")
-    private Double mobileCpc;
+    private BigDecimal mobileCpc;
 
     @Field(value = "mcpm")
-    private Double mobileCpm;
+    private BigDecimal mobileCpm;
 
     @Field(value = "mcs")
     private Double mobileConversion;
@@ -127,27 +128,27 @@ public class AccountReportEntity implements Comparable<AccountReportEntity>{
         this.pcCtr = pcCtr;
     }
 
-    public Double getPcCost() {
+    public BigDecimal getPcCost() {
         return pcCost;
     }
 
-    public void setPcCost(Double pcCost) {
+    public void setPcCost(BigDecimal pcCost) {
         this.pcCost = pcCost;
     }
 
-    public Double getPcCpc() {
+    public BigDecimal getPcCpc() {
         return pcCpc;
     }
 
-    public void setPcCpc(Double pcCpc) {
+    public void setPcCpc(BigDecimal pcCpc) {
         this.pcCpc = pcCpc;
     }
 
-    public Double getPcCpm() {
+    public BigDecimal getPcCpm() {
         return pcCpm;
     }
 
-    public void setPcCpm(Double pcCpm) {
+    public void setPcCpm(BigDecimal pcCpm) {
         this.pcCpm = pcCpm;
     }
 
@@ -183,27 +184,27 @@ public class AccountReportEntity implements Comparable<AccountReportEntity>{
         this.mobileCtr = mobileCtr;
     }
 
-    public Double getMobileCost() {
+    public BigDecimal getMobileCost() {
         return mobileCost;
     }
 
-    public void setMobileCost(Double mobileCost) {
+    public void setMobileCost(BigDecimal mobileCost) {
         this.mobileCost = mobileCost;
     }
 
-    public Double getMobileCpc() {
+    public BigDecimal getMobileCpc() {
         return mobileCpc;
     }
 
-    public void setMobileCpc(Double mobileCpc) {
+    public void setMobileCpc(BigDecimal mobileCpc) {
         this.mobileCpc = mobileCpc;
     }
 
-    public Double getMobileCpm() {
+    public BigDecimal getMobileCpm() {
         return mobileCpm;
     }
 
-    public void setMobileCpm(Double mobileCpm) {
+    public void setMobileCpm(BigDecimal mobileCpm) {
         this.mobileCpm = mobileCpm;
     }
